+++ example/example_lock.dart
@@ -1,0 +1,25 @@
+import 'package:multitasking/multitasking.dart';
+import 'package:multitasking/synchronization/binary_semaphore.dart';
+
+Future<void> main(List<String> args) async {
+  final sem = BinarySemaphore();
+  final tasks = <AnyTask>[];
+  for (var i = 0; i < 3; i++) {
+    final task = Task.run(() async {
+      await sem.lock(() async {
+        _message('Enter');
+        await Future<void>.delayed(Duration(milliseconds: 100));
+        _message('Leave');
+      });
+    });
+
+    tasks.add(task);
+  }
+
+  await Task.waitAll(tasks);
+}
+
+void _message(String text) {
+  final task = Task.current.name ?? '${Task.current}';
+  print('$task: $text');
+}