+++ example/example_task_cancel_network.dart
@@ -59,7 +59,7 @@ Future<void> main() async {
   for (final task in tasks) {
     print('-' * 40);
     print('${task.toString()}: ${task.state.name}');
-    if (task.state == TaskState.completed) {
+    if (task.isCompleted) {
       final value = await task;
       final text = value;
       final length = text.length < 80 ? text.length : 80;